feat: refactor code #9
