+++ src/components/index/Movie.tsx
@@ -1,6 +1,6 @@
 import { h } from 'preact'
 
-import { Button } from './ui/Button'
+import { Button } from '../ui/Button'
 
 interface Props {
   imageSrc: string
