+++ src/components/movies/ui/MovieCard.tsx
