+++ src/pages/movies.astro
@@ -1,5 +1,5 @@
 ---
-import { MoviesList } from '../components/MoviesList'
+import { MoviesList } from '../components/movies/MoviesList'
 import MainLayout from '../layouts/MainLayout.astro'
 import { MovieType } from '../utils/types'
 import { getDBMovies, getMoviesData } from '../utils/getMovies'
