--- src/utils/pickClientMovie.ts
@@ -1,8 +1,0 @@
-export const pickMovie = async (pick: string, against: string) => {
-  const res = await fetch('/api/pick-movie', {
-    method: 'POST',
-    body: JSON.stringify({ pick, against }),
-  })
-
-  return res.json()
-}
