+++ src/utils/utils.ts
@@ -1,3 +1,5 @@
+import { type MovieType } from './types'
+
 export const getRandomPage = (maxPage = 999, skip?: Set<number>) => {
   const result = Math.floor(Math.random() * (maxPage + 1))
 
@@ -6,4 +8,34 @@ export const getRandomPage = (maxPage = 999, skip?: Set<number>) => {
   }
 
   return result
+}
+
+// export const preloadImage = (src: string) => {
+//   new Image().src = src
+// }
+
+export const getDataFromImageCaption = (
+  movie: MovieType,
+): { cast?: string; title: string; alternativeTitle?: string } => {
+  const caption = movie.primaryImage.caption.plainText
+  const title = movie.titleText.text
+
+  if (!caption) {
+    return { title }
+  }
+
+  const indexOfTitle = caption.indexOf(title)
+
+  if (indexOfTitle === -1) {
+    return { title: caption, alternativeTitle: title }
+  }
+
+  if (indexOfTitle === 0) {
+    return { title: caption }
+  }
+
+  const cast = caption.slice(0, indexOfTitle - 4)
+  const titleWithYear = caption.slice(indexOfTitle)
+
+  return { cast, title: titleWithYear }
 }
